Setup CI and engine update workflow
Multiple different things are happening here:

1. Checked-in pre-compiled engines replaced by the download logic.

Engine hash is stored in checked-in `.versions/engine` file and is
downloaded in `prepare` script. Pre-compiled engines are still part of
published npm package but they are no longer checked into the
repository. That also removes the need for LFS setup.

2. CI workflow for automated testing is setup

It is pretty much e2e test from prisma/prisma moved into this repository
instead. Build example app and runs simple integration test on both iOS
and Android emulator. It also fixes a bunch of linting failures present
in the repo.

3. CI workflow for automated updates of client, engines, and automated
   publishing of the package.

Polls npm on cron, if the update found on either `prisma@dev` or
`prisma@latest`, pulls in the new version, does update on the separate
temporary branch and runs aforementioned test on it. If the test is
succsful, branch is merged back into main and new `@prisma/react-native`
package is published with an identical version and npm tag.

Close prisma/team-orm#1106