Ensure that test actually fails when things are broken
Test could never actually fail, because the only thing it doess is
checks that `test_indicator` element exists and that element was shown
regardless of test result. Fixed that by giving different test ids to
success and failure indicators.

At the moment, tests deliberately fail because custom databases are not
supported.